requirements.txt
biopython==1.85
torch==2.6.0
transformers==4.48.2
tokenizers==0.21.0
MDAnalysis==2.8.0
python-dotenv==1.0.1
torch_geometric==2.6.1
tqdm==4.66.5
fair-esm
scikit-learn
matplotlib